JEN-1192  Add possibility of specifying 'custom' docker

+++ cloud/jenkins/pxc_operator_minikube.groovy
@@ -27,13 +27,26 @@ void runTest(String TEST_NAME) {
             echo Skip $TEST_NAME test
         else
             cd ./source
-            export IMAGE=perconalab/percona-xtradb-cluster-operator:${env.GIT_BRANCH}
+            if [ -n "${PXC_OPERATOR_IMAGE}" ]; then
+                export IMAGE=${PXC_OPERATOR_IMAGE}
+            else
+                export IMAGE=perconalab/percona-xtradb-cluster-operator:${env.GIT_BRANCH}
+            fi
+
             if [ -n "${IMAGE_PXC}" ]; then
                 export IMAGE_PXC=${IMAGE_PXC}
             fi
 
             if [ -n "${IMAGE_PROXY}" ]; then
                 export IMAGE_PROXY=${IMAGE_PROXY}
+            fi
+
+            if [ -n "${IMAGE_BACKUP}" ]; then
+                export IMAGE_BACKUP=${IMAGE_BACKUP}
+            fi
+
+            if [ -n "${IMAGE_PMM}" ]; then
+                export IMAGE_PMM=${IMAGE_PMM}
             fi
 
             ./e2e-tests/$TEST_NAME/run
@@ -67,12 +80,24 @@ pipeline {
             name: 'GIT_REPO')
         string(
             defaultValue: '',
-            description: 'PXC image',
+            description: 'Operator image: perconalab/percona-xtradb-cluster-operator:master',
+            name: 'PXC_OPERATOR_IMAGE')
+        string(
+            defaultValue: '',
+            description: 'PXC image: perconalab/percona-xtradb-cluster-operator:master-pxc',
             name: 'IMAGE_PXC')
         string(
             defaultValue: '',
-            description: 'PXC proxy image',
+            description: 'PXC proxy image: perconalab/percona-xtradb-cluster-operator:master-proxysql',
             name: 'IMAGE_PROXY')
+        string(
+            defaultValue: '',
+            description: 'Backup image: perconalab/percona-server-mongodb-operator:master-backup',
+            name: 'IMAGE_BACKUP')
+        string(
+            defaultValue: '',
+            description: 'PMM image: perconalab/percona-server-mongodb-operator:master-pmm',
+            name: 'IMAGE_PMM')
     }
     agent {
          label 'micro-amazon' 
@@ -108,14 +133,18 @@ pipeline {
                 unstash "sourceFILES"
                 withCredentials([usernamePassword(credentialsId: 'hub.docker.com', passwordVariable: 'PASS', usernameVariable: 'USER')]) {
                     sh '''
-                        cd ./source/
-                        sg docker -c "
-                            docker login -u '${USER}' -p '${PASS}'
-                            export IMAGE=perconalab/percona-xtradb-cluster-operator:$GIT_BRANCH
-                            ./e2e-tests/build
-                            docker logout
-                        "
-                        sudo rm -rf ./build
+                        if [ -n "${PXC_OPERATOR_IMAGE}" ]; then
+                            echo "SKIP: Build is not needed, PXC operator image was set!"
+                        else
+                            cd ./source/
+                            sg docker -c "
+                                docker login -u '${USER}' -p '${PASS}'
+                                export IMAGE=perconalab/percona-xtradb-cluster-operator:$GIT_BRANCH
+                                ./e2e-tests/build
+                                docker logout
+                            "
+                            sudo rm -rf ./build
+                        fi
                     '''
                 }
             }
@@ -137,7 +166,7 @@ pipeline {
 
                         sudo curl -Lo /usr/local/bin/minikube https://storage.googleapis.com/minikube/releases/latest/minikube-linux-amd64
                         sudo chmod +x /usr/local/bin/minikube
-                        minikube start --vm-driver=virtualbox --dns-domain=percona.com --memory=4096 --cpus=3 --kubernetes-version v1.12.8
+                        minikube start --vm-driver=none --dns-domain=percona.com --memory=4096 --cpus=3 --kubernetes-version v1.12.8
                     '''
                     
                     unstash "sourceFILES"
